+++ src/lib/pages/proposal/store-code/index.tsx
@@ -490,10 +490,18 @@ const StoreCodeProposal = () => {
                   label={PROPOSAL_STORE_CODE_TEXT.builderLabel}
                   labelBgColor="background.main"
                   variant="floating"
+                  helperText={PROPOSAL_STORE_CODE_TEXT.builderHelperText}
+                  // Builder is a docker image, can be tagged, digested, or both
                   rules={{
                     required: PROPOSAL_STORE_CODE_TEXT.builderRequired,
+                    pattern: PROPOSAL_STORE_CODE_TEXT.builderPattern,
                   }}
-                  error={errors.builder?.message}
+                  error={
+                    builder &&
+                    !builder.match(PROPOSAL_STORE_CODE_TEXT.builderPattern)
+                      ? PROPOSAL_STORE_CODE_TEXT.builderError
+                      : errors.builder?.message
+                  }
                 />
 
                 {/* Source  */}
@@ -507,11 +515,12 @@ const StoreCodeProposal = () => {
                   helperText={PROPOSAL_STORE_CODE_TEXT.sourceHelperText}
                   rules={{
                     required: PROPOSAL_STORE_CODE_TEXT.sourceRequired,
-                    pattern: /^(?!:)[a-zA-Z0-9+.-]+:/,
+                    pattern: PROPOSAL_STORE_CODE_TEXT.sourcePattern,
                   }}
                   error={
                     // Source should be absolute url or absolute path
-                    source && !source.match(/^(?!:)[a-zA-Z0-9+.-]+:/)
+                    source &&
+                    !source.match(PROPOSAL_STORE_CODE_TEXT.sourcePattern)
                       ? PROPOSAL_STORE_CODE_TEXT.sourceHelperText
                       : errors.source?.message
                   }